Crate flo_curves

flo_curves

flo_curves is a library of routines for inspecting and manipulating curves, with a focus on cubic Bézier curves. In this library, you’ll find routines for computing points on curves, performing collision detection between curves and lines or other curves, all the way up to routines for combining paths made up of multiple curves.

Anyone doing any work with Bézier curves will likely find something in this library that is of use, but its range of functions makes it particularly useful for collision detection or performing path arithmetic.

A set of curve and coordinate types are provided by the library, as well as a set of traits that can be implemented on any types with suitable properties. Implementing these traits makes it possible to add the extra features of this library to any existing code that has its own way of representing coordinates, curves or paths.

flo_curves was built as a support library for flowbetween, an animation tool I’m working on.

Re-exports

pub use self::geo::*;

Modules

arc

Describing circular arcs

bezier

Routines for describing, querying and manipulating Bezier curves

debug

geo

Traits for basic geometric definitions

line

Manipulating and describing lines

Traits

BezierCurve

Trait implemented by things representing a cubic bezier curve

BezierCurveFactory

Trait implemented by bezier curves that can create new versions of themselves

Line

Represents a straight line