Function flo_curves::bezier::solve_curve_for_t_along_axis

pub fn solve_curve_for_t_along_axis<C: BezierCurve>(
    curve: &C,
    point: &C::Point,
    accuracy: f64
) -> Option<f64>

Searches along the x or y axis for a point within accuracy units of the curve, returning the t value of that point

This can be used to find points on the curve that are close to points that are not but does not find the closest point and it’s possible to construct a curve such that a nearby point is not nearby along either the x or the y axis. Therefore, this is best used with accuracy values that are small compared to the length of the curve and with points far away from inflection points or cusps.

This is often ‘good enough’ to find a point close to where a user has clicked along a curve, for example, but as it effectively ray-casts along the x and y axes to do so is not suitable as a general-purpose way of finding the closest point on a curve to another point.

Note that curve_intersects_ray() can be used to find points on a curve along any direction rather than solely along the axis.