Function flo_curves::bezier::path::path_intersect

pub fn path_intersect<P1: BezierPath, P2: BezierPath, POut: BezierPathFactory>(
    path1: &Vec<P1>,
    path2: &Vec<P2>,
    accuracy: f64
) -> Vec<POut>where
    P1::Point: Coordinate + Coordinate2D,
    P2: BezierPath<Point = P1::Point>,
    POut: BezierPathFactory<Point = P1::Point>,

Generates the path formed by intersecting two sets of paths

Each of the two paths passed into this function is assumed not to overlap themselves. IE, this does not perform self-intersection on either path1 or path2. This provides both a performance optimisation and finer control over how self-intersecting paths are handled. See path_remove_interior_points() and path_remove_overlapped_points() for a way to eliminate overlaps.

The input vectors represent the external edges of the path to intersect (a single BezierPath cannot have any holes in it, so a set of them effectively represents a path intended to be rendered with an even-odd winding rule)