Struct flo_curves::bezier::path::GraphPath

pub struct GraphPath<Point, Label> { /* private fields */ }

A graph path is a path where each point can have more than one connected edge. Edges are categorized into interior and exterior edges depending on if they are on the outside or the inside of the combined shape.

Implementations

impl<Point: Coordinate + Coordinate2D, Label> GraphPath<Point, Label>

pub fn following_edge_ref(&self, edge_ref: GraphEdgeRef) -> GraphEdgeRef

Given an edge ref, returns the edge ref that follows it

impl<Point: Coordinate + Coordinate2D, Label: Copy> GraphPath<Point, Label>

pub fn ray_collisions<L: Line<Point = Point>>(
&self,
ray: &L
) -> Vec<(GraphRayCollision, f64, f64, Point)>

Finds all collisions between a ray and this path

The return value is a tuple of (collision, curve_t, line_t, position)

impl<Point: Coordinate + Coordinate2D, Label: Copy> GraphPath<Point, Label>

pub fn combine_overlapping_points(&mut self, accuracy: f64) -> bool

Finds any points that have approximately the same coordinates and combines them

Accuracy indicates the maximum difference in the x or y coordinate for two points to be considered the same.

impl<Point: Coordinate + Coordinate2D, Label: Copy> GraphPath<Point, Label>

pub fn new() -> GraphPath<Point, Label>

Creates a new graph path with no points

pub fn from_path<P: BezierPath<Point = Point>>(
path: &P,
label: Label
) -> GraphPath<Point, Label>

Creates a graph path from a bezier path

pub fn from_merged_paths<'a, P: 'a + BezierPath<Point = Point>, PathIter: IntoIterator<Item = (&'a P, Label)>>(
paths: PathIter
) -> GraphPath<Point, Label>

Creates a new graph path by merging (not colliding) a set of paths with their labels

pub fn num_points(&self) -> usize

Returns the number of points in this graph. Points are numbered from 0 to this value.

pub fn all_edges(
&self
) -> impl '_ + Iterator<Item = GraphEdge<'_, Point, Label>>

Returns an iterator of all edges in this graph

pub fn all_edge_refs(&self) -> impl '_ + Iterator<Item = GraphEdgeRef>

Returns an iterator of all the edges in this graph, as references

pub fn edges_for_point(
&self,
point_num: usize
) -> impl '_ + Iterator<Item = GraphEdge<'_, Point, Label>>

Returns an iterator of the edges that leave a particular point

Edges are directional: this will provide the edges that leave the supplied point

pub fn edge_refs_for_point(
&self,
point_num: usize
) -> impl Iterator<Item = GraphEdgeRef>

Returns the edge refs for a particular point

pub fn point_position(&self, point_num: usize) -> Point

Returns the position of a particular point

pub fn reverse_edges_for_point(
&self,
point_num: usize
) -> impl '_ + Iterator<Item = GraphEdge<'_, Point, Label>>

Returns an iterator of the edges that arrive at a particular point

Edges are directional: this will provide the edges that connect to the supplied point

pub fn merge(
self,
merge_path: GraphPath<Point, Label>
) -> GraphPath<Point, Label>

Merges in another path

This adds the edges in the new path to this path without considering if they are internal or external

pub fn collide_or_merge(
self,
collide_path: GraphPath<Point, Label>,
accuracy: f64
) -> CollidedGraphPath<Point, Label>

Collides this path against another, generating a merged path

Anywhere this graph intersects the second graph, a point with two edges will be generated. All edges will be left as interior or exterior depending on how they’re set on the graph they originate from.

Working out the collision points is the first step to performing path arithmetic: the resulting graph can be altered to specify edge types - knowing if an edge is an interior or exterior edge makes it possible to tell the difference between a hole cut into a shape and an intersection.

Unlike collide(), this will indicate if any collisions were detected or if the two paths merged without collisions

pub fn collide(
self,
collide_path: GraphPath<Point, Label>,
accuracy: f64
) -> GraphPath<Point, Label>

Collides this path against another, generating a merged path

Anywhere this graph intersects the second graph, a point with two edges will be generated. All edges will be left as interior or exterior depending on how they’re set on the graph they originate from.

Working out the collision points is the first step to performing path arithmetic: the resulting graph can be altered to specify edge types - knowing if an edge is an interior or exterior edge makes it possible to tell the difference between a hole cut into a shape and an intersection.

pub fn round(&mut self, accuracy: f64)

Rounds all of the points in this path to a particular accuracy level

pub fn self_collide(&mut self, accuracy: f64)

Finds any collisions between existing points in the graph path

pub fn get_edge(&self, edge: GraphEdgeRef) -> GraphEdge<'_, Point, Label>

Returns the GraphEdge for an edgeref

pub fn set_edge_kind(&mut self, edge: GraphEdgeRef, new_type: GraphPathEdgeKind)

Sets the kind of a single edge

pub fn set_edge_label(&mut self, edge: GraphEdgeRef, new_label: Label)

Sets the label of a single edge

pub fn edge_kind(&self, edge: GraphEdgeRef) -> GraphPathEdgeKind

Returns the type of the edge pointed to by an edgeref

pub fn edge_label(&self, edge: GraphEdgeRef) -> Label

Returns the label of the edge pointed to by an edgeref

pub fn reset_edge_kinds(&mut self)

Resets the edge kinds in this path by setting them all to uncategorised

pub fn set_edge_kind_connected(
&mut self,
edge: GraphEdgeRef,
kind: GraphPathEdgeKind
)

Sets the kind of an edge and any connected edge where there are no intersections (only one edge)

pub fn heal_exterior_gaps(&mut self) -> bool

Finds any gaps in the edges marked as exterior and attempts to ‘heal’ them by finding a route to another part of the path with a missing edge

Returns true if all the gaps that were found were ‘healed’

pub fn exterior_paths<POut: BezierPathFactory<Point = Point>>(
&self
) -> Vec<POut>

Finds the exterior edges and turns them into a series of paths

impl<Point: Coordinate + Coordinate2D> GraphPath<Point, PathLabel>

pub fn edge_collision_count(
&self,
target_edge: GraphEdgeRef,
t: f64
) -> Option<i64>

Computes the collision count for a point along an edge in the graph

The result is ‘None’ if the point described is at an intersection or if a ray cast at the specified point does not intersect the edge.

The collision count is uncertain if it occurs on two overlapping edges, as any ordering of the edges is valid. set_edge_kinds_by_ray_casting() has rules for this so that crossing counts are consistent when entering and leaving a shape. This call uses simpler rules and might switch which shape is hit first when two shapes have edges is common.

pub fn set_edge_kinds_by_ray_casting<FnIsInside: Fn(&SmallVec<[i32; 8]>) -> bool>(
&mut self,
is_inside: FnIsInside
)

Sets the edge kinds by performing ray casting

The function passed in to this method takes two parameters: these are the number of times edges have been crossed in path 1 and path 2. It should return true if this number of crossings represents a point inside the final shape, or false if it represents a point outside of the shape.

Path crossings are processed in the order they’re hit by the ray, with some exceptions:

If a ray hits an intersection or hits very close to an intersection, the order is arbitrary and the edge kinds are not updated for that intersection (but are updated later on).

If a ray crosses a set of overlapping edges, the order that the edges are crossed in depends on whether or not the ray is considered to be entering or leaving the ‘inner’ of the two shapes. If it’s found to be entering the shape, the ray will hit the edge belonging to the second shape first, and if it’s leaving it will hit the edge belonging to the first shape first. This ensures that the behaviour is consistent when the ray’s direction is reversed.

impl<Point: Coordinate + Coordinate2D> GraphPath<Point, PathLabel>

pub fn set_exterior_by_intersecting(&mut self)

Given a labelled graph path, marks exterior edges by intersecting PathSource::Path1 and PathSource::Path2

impl<Point: Coordinate + Coordinate2D> GraphPath<Point, PathLabel>

pub fn set_exterior_by_adding(&mut self)

Given a labelled graph path, marks exterior edges by adding PathSource::Path1 and PathSource::Path2

This is used to implement the path_add() arithmetic function, and is generally called after using collide() to combine the two input paths.

pub fn set_exterior_by_removing_interior_points(&mut self)

Given a path that intersects itself (ie, only contains SourcePath::Path1), discovers the ‘true’ exterior edge.

impl<Point: Coordinate + Coordinate2D> GraphPath<Point, PathLabel>

pub fn set_exterior_by_subtracting(&mut self)

Given a labelled graph path, marks exterior edges by subtracting PathSource::Path2 from PathSource::Path1