Trait flo_curves::bezier::path::BezierPath

pub trait BezierPath: Geo + Clone + Sized {
    type PointIter: Iterator<Item = (Self::Point, Self::Point, Self::Point)>;

    fn start_point(&self) -> Self::Point;
    fn points(&self) -> Self::PointIter;

    fn bounding_box<Bounds: BoundingBox<Point = Self::Point>>(&self) -> Bounds { ... }
    fn fast_bounding_box<Bounds: BoundingBox<Point = Self::Point>>(
        &self
    ) -> Bounds { ... }
    fn to_curves<Curve: BezierCurveFactory<Point = Self::Point>>(
        &self
    ) -> Vec<Curve> { ... }
    fn reversed<POut: BezierPathFactory<Point = Self::Point>>(&self) -> POut { ... }
}

Trait representing a path made out of bezier sections

Required Associated Types

type PointIter: Iterator<Item = (Self::Point, Self::Point, Self::Point)>

Type of an iterator over the points in this curve. This tuple contains the points ordered as a hull: ie, two control points followed by a point on the curve

Required Methods

fn start_point(&self) -> Self::Point

Retrieves the initial point of this path

fn points(&self) -> Self::PointIter

Retrieves an iterator over the points in this path

Provided Methods

fn bounding_box<Bounds: BoundingBox<Point = Self::Point>>(&self) -> Bounds

Finds the bounds of this path

fn fast_bounding_box<Bounds: BoundingBox<Point = Self::Point>>(&self) -> Bounds

Finds a loose bounding box for this path (more quickly than bounding_box)

This will contain the path but might not be tightly aligned to the curves

fn to_curves<Curve: BezierCurveFactory<Point = Self::Point>>(
&self
) -> Vec<Curve>

Changes this path into a set of bezier curves

fn reversed<POut: BezierPathFactory<Point = Self::Point>>(&self) -> POut

Creates a reversed version of this path