Trait flo_curves::bezier::path::BezierPathFactory

pub trait BezierPathFactory: BezierPath {
    fn from_points<FromIter: IntoIterator<Item = (Self::Point, Self::Point, Self::Point)>>(
        start_point: Self::Point,
        points: FromIter
    ) -> Self;

    fn from_path<FromPath: BezierPath<Point = Self::Point>>(
        path: &FromPath
    ) -> Self { ... }
}

Trait implemented by types that can construct new bezier paths

Required Methods

fn from_points<FromIter: IntoIterator<Item = (Self::Point, Self::Point, Self::Point)>>(
start_point: Self::Point,
points: FromIter
) -> Self

Creates a new instance of this path from a set of points

Provided Methods

fn from_path<FromPath: BezierPath<Point = Self::Point>>(path: &FromPath) -> Self

Creates a new instance of this path from the points in another path