Struct flo_curves::bezier::CurveSection

pub struct CurveSection<'a, C: 'a + BezierCurve> { /* private fields */ }

Represents a subsection of a bezier curve

Implementations

impl<'a, C: 'a + BezierCurve> CurveSection<'a, C>

pub fn new(curve: &'a C, t_min: f64, t_max: f64) -> CurveSection<'a, C>

Creates a new curve section from a region of another bezier curve

pub fn t_for_t(&self, t: f64) -> f64

Returns the t value on the full curve for a t value on the section

pub fn is_tiny(&self) -> bool

Returns true if this section is so small as to represent a point

pub fn subsection(&self, t_min: f64, t_max: f64) -> CurveSection<'a, C>

Creates a sub-section from this curve section (dividing it further)

Unlike calling section, this keeps the same type and avoids the need for recursive recalculation for things like the control points. This means that original_curve_t_values will return the coordinates for the same original curve as the curve that this subsection was created from.

pub fn original_curve_t_values(&self) -> (f64, f64)

Returns the original t values (t_min, t_max) that this section was created from

pub fn section_t_for_original_t(&self, t: f64) -> f64

Given a ‘t’ value on the original curve, returns the equivalent value on this section