Trait flo_curves::bezier::BezierCurve

pub trait BezierCurve: Geo + Clone + Sized {

    fn start_point(&self) -> Self::Point;
    fn end_point(&self) -> Self::Point;
    fn control_points(&self) -> (Self::Point, Self::Point);

    fn reverse<Curve: BezierCurveFactory<Point = Self::Point>>(self) -> Curve { ... }
    fn point_at_pos(&self, t: f64) -> Self::Point { ... }
    fn t_for_point(&self, point: &Self::Point) -> Option<f64> { ... }
    fn subdivide<Curve: BezierCurveFactory<Point = Self::Point>>(
        &self,
        t: f64
    ) -> (Curve, Curve) { ... }
    fn bounding_box<Bounds: BoundingBox<Point = Self::Point>>(&self) -> Bounds { ... }
    fn fast_bounding_box<Bounds: BoundingBox<Point = Self::Point>>(
        &self
    ) -> Bounds { ... }
    fn search_with_bounds<MatchFn: Fn(Self::Point, Self::Point) -> bool>(
        &self,
        max_error: f64,
        match_fn: MatchFn
    ) -> Vec<f64> { ... }
    fn find_extremities(&self) -> Vec<f64> { ... }
    fn estimate_length(&self) -> f64 { ... }
    fn section(&self, t_min: f64, t_max: f64) -> CurveSection<'_, Self> { ... }

}

Trait implemented by things representing a cubic bezier curve

Required Methods

fn start_point(&self) -> Self::Point

The start point of this curve

fn end_point(&self) -> Self::Point

The end point of this curve

fn control_points(&self) -> (Self::Point, Self::Point)

The control points in this curve

Provided Methods

fn reverse<Curve: BezierCurveFactory<Point = Self::Point>>(self) -> Curve

Reverses the direction of this curve

fn point_at_pos(&self, t: f64) -> Self::Point

Given a value t from 0 to 1, returns a point on this curve

fn t_for_point(&self, point: &Self::Point) -> Option<f64>

Given a point that is on or very close to the curve, returns the t value where the point can be found (or None if the point is not very close to the curve)

fn subdivide<Curve: BezierCurveFactory<Point = Self::Point>>(
&self,
t: f64
) -> (Curve, Curve)

Given a value t from 0 to 1, finds a point on this curve and subdivides it, returning the two resulting curves

fn bounding_box<Bounds: BoundingBox<Point = Self::Point>>(&self) -> Bounds

Computes the bounds of this bezier curve

fn fast_bounding_box<Bounds: BoundingBox<Point = Self::Point>>(&self) -> Bounds

Faster but less accurate bounding box for a curve

This will produce a bounding box that contains the curve but which may be larger than necessary

fn search_with_bounds<MatchFn: Fn(Self::Point, Self::Point) -> bool>(
&self,
max_error: f64,
match_fn: MatchFn
) -> Vec<f64>

Given a function that determines if a searched-for point is within a bounding box, searches the curve for the t values for the corresponding points

fn find_extremities(&self) -> Vec<f64>

Finds the t values where this curve has extremities

fn estimate_length(&self) -> f64

Attempts to estimate the length of this curve

fn section(&self, t_min: f64, t_max: f64) -> CurveSection<'_, Self>

Create a section from this curve. Consider calling subsection for curves that are already CurveSections.