Trait flo_curves::geo::BoundingBox

pub trait BoundingBox: Geo + Sized {
    fn from_min_max(min: Self::Point, max: Self::Point) -> Self;
    fn min(&self) -> Self::Point;
    fn max(&self) -> Self::Point;

    fn bounds_for_points<PointIter: IntoIterator<Item = Self::Point>>(
        points: PointIter
    ) -> Self { ... }
    fn empty() -> Self { ... }
    fn is_empty(&self) -> bool { ... }
    fn union_bounds(self, target: Self) -> Self { ... }
    fn overlaps(&self, target: &Self) -> bool { ... }
}

Trait implemented by things representing axis-aligned bounding boxes

Required Methods

fn from_min_max(min: Self::Point, max: Self::Point) -> Self

Returns a bounding box with the specified minimum and maximum coordinates

fn min(&self) -> Self::Point

Returns the minimum point of this bounding box

fn max(&self) -> Self::Point

Returns the maximum point of this bounding box

Provided Methods

fn bounds_for_points<PointIter: IntoIterator<Item = Self::Point>>(
points: PointIter
) -> Self

Returns a bounding box containing the specified points

fn empty() -> Self

Returns an empty bounding box

fn is_empty(&self) -> bool

True if this bounding box is empty

fn union_bounds(self, target: Self) -> Self

Creates the union of this and another bounding box

fn overlaps(&self, target: &Self) -> bool

Returns true if this bounding box overlaps another