Trait flo_curves::geo::Coordinate

pub trait Coordinate: Sized + Copy + Add<Self, Output = Self> + Mul<f64, Output = Self> + Sub<Self, Output = Self> + PartialEq {

    fn from_components(components: &[f64]) -> Self;
    fn origin() -> Self;
    fn len() -> usize;
    fn get(&self, index: usize) -> f64;
    fn from_biggest_components(p1: Self, p2: Self) -> Self;
    fn from_smallest_components(p1: Self, p2: Self) -> Self;

    fn distance_to(&self, target: &Self) -> f64 { ... }
    fn dot(&self, target: &Self) -> f64 { ... }
    fn magnitude(&self) -> f64 { ... }
    fn to_unit_vector(&self) -> Self { ... }
    fn is_nan(&self) -> bool { ... }
    fn round(self, accuracy: f64) -> Self { ... }
    fn is_near_to(&self, other: &Self, max_distance: f64) -> bool { ... }
    fn smooth(points: &[Self], weights: &[f64]) -> Vec<Self> { ... }

}

Represents a value that can be used as a coordinate in a bezier curve

Required Methods

fn from_components(components: &[f64]) -> Self

Creates a new coordinate from the specified set of components

fn origin() -> Self

Returns the origin coordinate

fn len() -> usize

The number of components in this coordinate

fn get(&self, index: usize) -> f64

Retrieves the component at the specified index

fn from_biggest_components(p1: Self, p2: Self) -> Self

Returns a point made up of the biggest components of the two points

fn from_smallest_components(p1: Self, p2: Self) -> Self

Returns a point made up of the smallest components of the two points

Provided Methods

fn distance_to(&self, target: &Self) -> f64

Computes the distance between this coordinate and another of the same type

fn dot(&self, target: &Self) -> f64

Computes the dot product for this vector along with another vector

fn magnitude(&self) -> f64

Computes the magnitude of this vector

fn to_unit_vector(&self) -> Self

Treating this as a vector, returns a unit vector in the same direction

fn is_nan(&self) -> bool

Returns true if this coordinate has a NaN component

fn round(self, accuracy: f64) -> Self

Round this coordinate so that it is accurate to a certain precision

fn is_near_to(&self, other: &Self, max_distance: f64) -> bool

True if this point is within max_distance of another point

fn smooth(points: &[Self], weights: &[f64]) -> Vec<Self>

Generates a smoothed version of a set of coordinates, using the specified weights (weights should add up to 1.0).

A suggested set of weights might be ‘[0.25, 0.5, 0.25]’, which will slightly adjust each point according to its neighbours (the central weight is what’s applied to the ‘current’ point)