Function flo_curves::line::line_intersects_line

pub fn line_intersects_line<L: Line>(line1: &L, line2: &L) -> Option<L::Point>where
    L::Point: Coordinate2D,

Returns the point at which two lines intersect (if they intersect)

Only the 2-dimensional form is supported at the moment (lines are much less likely to intersect in higher dimensions)