Trait flo_curves::line::Line2D

pub trait Line2D {
    type Point: Coordinate + Coordinate2D;

    fn coefficients(&self) -> (f64, f64, f64);
    fn distance_to(&self, p: &Self::Point) -> f64;
    fn which_side(&self, p: &Self::Point) -> i8;
}

Trait implemented by a 2D line

Required Associated Types

type Point: Coordinate + Coordinate2D

Required Methods

fn coefficients(&self) -> (f64, f64, f64)

Returns the coefficients (a, b, c) for this line, such that ax+by+c = 0 for any point on the line and also such that a^2 + b^2 = 1

fn distance_to(&self, p: &Self::Point) -> f64

Returns the distance from a point to the closest point along this line

Note that this will project the line to infinity so this can return a distance to a point outside of the start or end point of the line. To determine if this has occurred, the pos_for_point() call can be used to determine the t value for the closest point: it will return a value in the range 0.0..1.0 if the closest point is within the line.

fn which_side(&self, p: &Self::Point) -> i8

Returns a value indicating which side of the line the specified point is on (+1, 0 or -1)